Fixes #5 Image response capabilities are added to reply functionality
This implements an image reply function which can be used later depending on the logic of the application. This commit allows the program to generate image based replies from loklak's visualization API's so that images can be sent. This has the potential to send geolocation based maps, markdown visualizations and piecharts which can be queried and images sent over.

Fixes https://github.com/loklak/asksusi_messenger/issues/5